Factor out volume-independent code from data viewer class
