+++ glue_vispy_viewers/scatter/scatter_viewer.py
